images/webapp/requirements.txt
Django==2.2.1
psycopg2==2.8
pytz==2018.9
djangorestframework==3.9.3
django-rest-swagger==2.2.0
dateutils==0.6.6
